% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{baseline}
\alias{baseline}
\title{Generates a baseline based on historical data}
\usage{
baseline(sales_data, promo_done = FALSE, sizeroll = 11,
  smoother = "mean", showgraph = FALSE)
}
\arguments{
\item{sales_data}{A vector containing historical sales data}

\item{promo_done}{A logical variable specifying if promotions are done for the product.}

\item{sizeroll}{An odd integer that determines the width of the rolling median
and rolling average to take.}

\item{smoother}{The smoother that should be considered. It can be "mean", "median" or "loess".}

\item{showgraph}{A logical value. If TRUE, returns the graph obtained from the analysis.}
}
\value{
A vector containing the historical values, a vector containing the baseline, 
 vector composed of binary variables that indicate whether an external impactor was detected 
 for a given period.

In option, the graph obtained in the process if *showgraph* == TRUE.
}
\description{
This function enables to generate a baseline based on historical
sales data and to plot the graphs linked to the different step of the
process to have a deep view on a given series. It eliminates the role of external impactors that have a significant
effect on the sales to obtain a smooth baseline.
}
\examples{
baseline() ###################### A COMPLETER ###############
}
\author{
Grandadam Patrik
}
