% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global.R
\name{contr.dummy}
\alias{contr.dummy}
\title{Returns a matrix of contrasts for the train.kknn.}
\usage{
contr.dummy(n, contrasts = TRUE)
}
\arguments{
\item{n}{A vector containing levels of a factor, or the number of levels.}

\item{contrasts}{A logical value indicating whether contrasts should be computed.}
}
\description{
Returns a matrix of contrasts for the train.kknn.
}
\examples{
contr.dummy(5)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
