% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global.R
\name{data.frame.dummy}
\alias{data.frame.dummy}
\title{Convierte toda la tabla a código dummy.}
\usage{
data.frame.dummy(DF, exclude = NULL)
}
\arguments{
\item{DF}{a data.frame.}

\item{exclude}{variables of data.frame exclude of conversion.}
}
\description{
Convierte toda la tabla a código dummy.
}
\examples{
data.frame.dummy(iris)

}
\author{
Diego Jimenez <diego.jimenezs@promidat.com>
}
