% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_boost_utils.R
\name{e_ada_evol_error}
\alias{e_ada_evol_error}
\title{Error Evolution}
\usage{
e_ada_evol_error(modelo, datos, label = "Iterations")
}
\arguments{
\item{modelo}{a adabag model.}

\item{datos}{a data.frame object.}

\item{label}{a label plot.}
}
\value{
echarts4r plot
}
\description{
Error Evolution
}
\examples{
model <- traineR::train.adabag(Species~., iris, mfinal = 20, coeflearn = 'Freund')
e_ada_evol_error(model, iris)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
