% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_boost_utils.R
\name{e_boost_importance}
\alias{e_boost_importance}
\title{Var importance Random Forest}
\usage{
e_boost_importance(modelo)
}
\arguments{
\item{modelo}{a adabag model.}
}
\value{
echarts4r plot
}
\description{
Var importance Random Forest
}
\examples{
model <- traineR::train.adabag(Species~., iris, mfinal = 20, coeflearn = 'Freund')
e_boost_importance(model)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
