% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_penalized_l_r_utils.R
\name{e_coeff_landa}
\alias{e_coeff_landa}
\title{Coefficients and lambda}
\usage{
e_coeff_landa(model, category, sel.lambda = NULL, label = "Log Lambda")
}
\arguments{
\item{model}{a glmnet model.}

\item{category}{a category of the variable to be predicted.}

\item{sel.lambda}{the selected lambda.}

\item{label}{a character specifying the title to use on selected lambda tooltip.}
}
\value{
echarts4r plot
}
\description{
Plot the coefficients and selected lambda of a glmnet model.
}
\examples{
modelo <- traineR::train.glmnet(Species~., iris)
e_coeff_landa(modelo, 'setosa', log(modelo$lambda[1]))

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
