% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global.R
\name{e_global_gauge}
\alias{e_global_gauge}
\title{Gauge Plot}
\usage{
e_global_gauge(
  value = 100,
  label = "Label",
  color1 = "#B5E391",
  color2 = "#90C468"
)
}
\arguments{
\item{value}{a number specifying the value of the graph.}

\item{label}{a character specifying the title to use on legend.}

\item{color1}{a color for the gauge.}

\item{color2}{a shadowColor for the gauge.}
}
\value{
echarts4r plot
}
\description{
Gauge Plot
}
\examples{
e_global_gauge(87, "Global Precision")

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
