% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global.R
\name{e_JS}
\alias{e_JS}
\title{Eval character vectors to JS code}
\usage{
e_JS(...)
}
\arguments{
\item{...}{character vectors to evaluate}
}
\description{
Eval character vectors to JS code
}
\examples{
e_JS('5 * 3')

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
