% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_regp_utils.R
\name{e_posib_lambda}
\alias{e_posib_lambda}
\title{Possible lambda}
\usage{
e_posib_lambda(
  cv.glm,
  labels = c("Valor Superior", "Valor Inferior", "lambda")
)
}
\arguments{
\item{cv.glm}{a cv.glmnet model.}

\item{labels}{a character vector of length 3 specifying the titles to use on legend.}
}
\value{
echarts4r plot
}
\description{
Possible lambda
}
\examples{
x         <- model.matrix(Species~., iris)[, -1]
y         <- iris[,'Species']
cv.glm    <- glmnet::cv.glmnet(x, y, standardize = TRUE, alpha = 1, family = 'multinomial')
e_posib_lambda(cv.glm)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
