% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_rndf_utils.R
\name{e_rf_error}
\alias{e_rf_error}
\title{Error Evolution}
\usage{
e_rf_error(model, label = "Trees")
}
\arguments{
\item{model}{a random forest model.}

\item{label}{a label plot.}
}
\value{
echarts4r plot
}
\description{
Error Evolution
}
\examples{
model <- traineR::train.randomForest(Species~., iris, mtry = 2, ntree = 20)
label <- "Trees"
e_rf_error(model, label)


}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
