% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_rndf_utils.R
\name{e_rndf_importance}
\alias{e_rndf_importance}
\title{Var importance Random Forest}
\usage{
e_rndf_importance(modelo, error = "MeanDecreaseAccuracy")
}
\arguments{
\item{modelo}{a random forest model.}

\item{error}{a character specifying the type of importance.}
}
\value{
echarts4r plot
}
\description{
Var importance Random Forest
}
\examples{
model <- traineR::train.randomForest(Species~., iris, mtry = 2, ntree = 20)
e_rndf_importance(model)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
