% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_xgb_utils.R
\name{e_xgb_importance}
\alias{e_xgb_importance}
\title{Var importance XGBoosting}
\usage{
e_xgb_importance(modelo, error = "Gain")
}
\arguments{
\item{modelo}{a random forest model.}

\item{error}{a character specifying the type of importance.}
}
\value{
echarts4r plot
}
\description{
Var importance XGBoosting
}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
