% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_predictor.R
\name{init_predictor}
\alias{init_predictor}
\title{This function will start predictoR}
\usage{
init_predictor()
}
\value{
Nothing
}
\description{
An interactive 'Shiny' application for data prediction.
}
\details{
Start predictoR

This starts the predictoR application on the user's local computer.
}
\examples{
 if(interactive()){
   init_predictor()
 }
}
\keyword{predictoR}
