% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_comparacion.R
\name{mod_comparacion_server}
\alias{mod_comparacion_server}
\title{comparacion Server Function
Update on load testing data}
\usage{
mod_comparacion_server(input, output, session, updateData)
}
\description{
comparacion Server Function
Update on load testing data
}
