% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ind_nuevos.R
\name{mod_ind_nuevos_server}
\alias{mod_ind_nuevos_server}
\title{ind_nuevos Server Function}
\usage{
mod_ind_nuevos_server(input, output, session, updateData, newCases)
}
\description{
Load Button Function
Load Button Function
Update Transform Table
Transform Button Function
Update Models Opcions
}
