% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poder_pred.R
\name{mod_poder_pred_server}
\alias{mod_poder_pred_server}
\title{poder_pred Server Function}
\usage{
mod_poder_pred_server(input, output, session, updateData)
}
\description{
Gráfico de Distribución Variable a Predecir 
Update on load testing data
}
