% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{my_theme}
\alias{my_theme}
\title{A great theme for graphs issued from the ggplot2 plackage}
\usage{
my_theme(base_size = 10, base_family = "sans")
}
\arguments{
\item{base_size}{The base size, no need to change it.}

\item{base_family}{The base_family, no need to change it}
}
\description{
This function enables to have a nice looking theme for ggplot graphs.
}
\author{
Grandadam Patrik
}
