% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{oos_simulation}
\alias{oos_simulation}
\title{Simulating the stock level and OOS probability}
\usage{
oos_simulation(forecast, sd, productions, stock, nsim = 1000)
}
\arguments{
\item{forecast}{A vector containing the forecasts of the future values}

\item{sd}{The standard deviation of the forecasts}

\item{productions}{A vector containing the future productions}

\item{stock}{The current level of stock}

\item{nsim}{The number of simulations to operate}
}
\description{
This function enables to calculate the expected stock level at the end of 
periods according to the current stock, the future forecasts, the productions. It is based on 
a Monte Carlo simulation that explores multiple demand scenarios.
}
\examples{
oos_simulation(
forecast = c(400, 380, 420, 560, 500, 480, 570, 600, 560, 590),
sd = 100,
productions = c(0, 0, 0, 2900, 0, 0, 0, 2000, 0, 0), 
stock = 1350,
nsim = 1000
)
}
\author{
Grandadam Patrik
}
