% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{percent_accuracy}
\alias{percent_accuracy}
\title{Forecasting accuracy as defined in the company}
\usage{
percent_accuracy(forecast, actuals)
}
\arguments{
\item{forecast}{The forecasts of the actuals that have been made previously.}

\item{actuals}{The actuals that are observed. It can be the values of the
testing set to evaluate the quality of a model, or the values that just have been observed
to estimate the real accuracy .}
}
\description{
This function enables to calculate the forecasting accuracy using
the same reporting measure as the company
}
\author{
Grandadam Patrik
}
