% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictoR.R
\docType{package}
\name{predictoR}
\alias{predictoR-package}
\alias{predictoR}
\title{Predictive Data Analysis System}
\description{
Perform a supervised data analysis on a database through a 'shiny' graphical interface. 
It includes methods such as K-Nearest Neighbors, Decision Trees, ADA Boosting, 
Extreme Gradient Boosting, Random Forest, Neural Networks, Deep Learning, 
Support Vector Machines and Bayesian Methods.
}
\details{
\tabular{ll}{
Package: \tab predictoR\cr
Type: \tab Package\cr
Version: \tab 4.0.5\cr
Date: \tab 2024-10-07\cr
License: \tab GPL (>=2)\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://promidat.website/}
  \item Report bugs at \url{https://github.com/PROMiDAT/predictoR/issues}
}

}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
}
\keyword{package}
