% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_comparacion_utils.R
\name{prob.values}
\alias{prob.values}
\title{Create Cut-Off Probability values.}
\usage{
prob.values(Score, Class, levels, category, step = -0.05)
}
\arguments{
\item{Score}{a data.frame object.}

\item{Class}{the column name to apply disjunctive code.}

\item{levels}{the column name to apply disjunctive code.}

\item{category}{a character value specifying the name of the category to apply the Cut-Off Probability.}

\item{step}{the step for the Cut-Off Probability.}
}
\description{
Create Cut-Off Probability values.
}
\author{
Joseline Quirós <joseline.quiros@promidat.com>
}
