% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_svm_utils.R
\name{voronoi_svm_plot}
\alias{voronoi_svm_plot}
\title{Voronoi Plot SVM}
\usage{
voronoi_svm_plot(datos, varpred, vars, kernel = "linear")
}
\arguments{
\item{datos}{a data.frame object.}

\item{varpred}{variable to predict.}

\item{vars}{predictor variables.}

\item{kernel}{the kernel used in training and predicting.}
}
\value{
plot
}
\description{
Voronoi Plot SVM
}
\examples{
voronoi_svm_plot(iris, "Species", c("Sepal.Length", "Sepal.Width"), "linear")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
