% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{predict_sales}
\alias{predict_sales}
\title{Creates sales forecast based on historical data}
\usage{
predict_sales(sales_data, frequency = 52, start = c(2014, 1),
  end = c(2019, 12), forecast_horizon = 52, size.te.test = 52,
  promo_done = FALSE, future_promotions = NA, criterion = "MAPE")
}
\arguments{
\item{sales_data}{A vector containing historical sales data}

\item{frequency}{A numerical value specifying the frequency of the seasonality}

\item{start}{A vector of length 2 with the date of the first observation.
It contains first the year and then the day/week/month according to your data.}

\item{end}{A vector of length 2 with the date of the last observation.
It contains first the year and then the day/week/month according to your data.}

\item{forecast_horizon}{An integer value specifying the number of observations to forecast.}

\item{size.te.test}{A integer value specifying the size of the training set}

\item{promo_done}{A logical variable specifying if promotions are done for the product.}

\item{future_promotions}{Optionnal: a vector composed on binary variables which are
equal to 1 when there will be a promotion in the forecasting horizon and to 0 otherwise.}

\item{criterion}{A string variable specifying the selection criterion that should be used to
select the model ("ME", "RMSE", "MAE", "MPE", "MAPE", "MASE", "ACF1", "Theil's U").}
}
\value{
A list containing the select model, the associated graphs, the predictions and the
confidence intervales, the accuracy measures and the same elements for all other considered models.
}
\description{
This function enables to create sales forecast based on historical
sales data. It considers a multitude of models and selects the model that had
the best performance on the testing set.
}
\examples{
predict_sales ###################### A COMPLETER ###############
}
\author{
Grandadam Patrik
}
