% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictoR.R
\docType{package}
\name{predictoR}
\alias{predictoR}
\title{Predictive Data Analysis System}
\description{
Perform a supervised data analysis on a database through a 'shiny' graphical interface. 
It includes methods such as K-Nearest Neighbors, Decision Trees, ADA Boosting, 
Extreme Gradient Boosting, Random Forest, Neural Networks, Deep Learning, 
Support Vector Machines and Bayesian Methods.
}
\details{
\tabular{ll}{
Package: \tab predictoR\cr
Type: \tab Package\cr
Version: \tab 3.0.0\cr
Date: \tab 2022-02-09\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
}
\keyword{package}
